+++ src/main/clojure/com/github/clojure_repl/intellij/configuration/factory/remote.clj
@@ -1,5 +1,6 @@
 (ns com.github.clojure-repl.intellij.configuration.factory.remote
   (:require
+   [clojure.java.io :as io]
    [com.github.clojure-repl.intellij.configuration.factory.base :as config.factory.base]
    [com.github.clojure-repl.intellij.db :as db]
    [com.github.ericdallo.clj4intellij.logger :as logger]
@@ -16,14 +17,25 @@
    [com.intellij.execution.process NopProcessHandler ProcessListener]
    [com.intellij.execution.runners ExecutionEnvironment]
    [com.intellij.openapi.project Project ProjectManager]
-   [com.intellij.ui IdeBorderFactory]))
+   [com.intellij.ui IdeBorderFactory]
+   [javax.swing JRadioButton JTextField]))
 
 (set! *warn-on-reflection* true)
 
 (def ^:private options-class ReplRemoteRunOptions)
 
+(defn ^:private read-port-file? []
+  (= (-> @db/db* :settings :mode) :file-config))
+
 (defn ^:private setup-process [project]
   (logger/info "Connecting to nREPL process...")
+  ;TODO: handle when file does not exist
+  (when (read-port-file?)
+    (let [base-path (.getBasePath ^Project project)
+          repl-file (io/file base-path ".nrepl-port")
+          port (slurp repl-file)]
+      (swap! db/db* assoc-in [:settings :nrepl-port] (parse-long port))))
+
   (let [handler (NopProcessHandler.)]
     (swap! config.factory.base/current-repl* assoc :handler handler)
     (.addProcessListener handler
@@ -32,28 +44,56 @@
                            (processWillTerminate [_ _ _] (config.factory.base/repl-disconnected))))
     handler))
 
+(defn manual? [editor]
+  (.isSelected ^JRadioButton (seesaw/select editor [:#manual])))
+
+(defn mode-id-key [repl-mode]
+  (->> (seesaw/selection repl-mode)
+       (seesaw/id-of)))
+
 (defn ^:private build-editor-view []
-  (mig/mig-panel
-   :border (IdeBorderFactory/createTitledBorder "nREPL connection")
-   :items [[(seesaw/label "Host") ""]
-           [(seesaw/text :id :nrepl-host
-                         :columns 20) "wrap"]
-           [(seesaw/label "Port") ""]
-           [(seesaw/text :id :nrepl-port
-                         :columns 8) "wrap"]
-           [(seesaw/label "Project") ""]
-           [(seesaw/combobox :id    :project
-                             :model (->> (ProjectManager/getInstance)
-                                         .getOpenProjects
-                                         (map #(.getName ^Project %)))) "wrap"]]))
+  (let [repl-mode-group (seesaw/button-group)
+        panel (mig/mig-panel
+                :border (IdeBorderFactory/createTitledBorder "NREPL connection")
+                :items [[(seesaw/radio :text "Manual"
+                                       :id :manual
+                                       :group repl-mode-group
+                                       :mnemonic \M
+                                       :selected? true) "wrap"]
+                        [(seesaw/label "Host") ""]
+                        [(seesaw/text :id :nrepl-host
+                                      :columns 20) "wrap"]
+                        [(seesaw/label "Port") ""]
+                        [(seesaw/text :id :nrepl-port
+                                      :columns 8) "wrap"]
+                        [(seesaw/radio :text "Read from repl file"
+                                       :id :repl-file
+                                       :group repl-mode-group
+                                       :mnemonic \R) "wrap"]
+                        [(seesaw/label "Project") ""]
+                        [(seesaw/combobox :id :project
+                                          :model (->> (ProjectManager/getInstance)
+                                                      .getOpenProjects
+                                                      (map #(.getName ^Project %)))) "wrap"]])]
+    (seesaw/listen repl-mode-group :action
+                   (fn [_e]
+                     (let [mode-key (mode-id-key repl-mode-group)
+                           manual? (= mode-key :manual)]
+                       (.setEnabled ^JTextField (seesaw/select panel [:#nrepl-host]) manual?)
+                       (.setEnabled ^JTextField (seesaw/select panel [:#nrepl-port]) manual?))))
+
+    panel))
 
 (set! *warn-on-reflection* false)
 (defn ^:private apply-editor-to [^RunConfigurationBase configuration-base]
   (let [editor @config.factory.base/editor-view*
         host (seesaw/text (seesaw/select editor [:#nrepl-host]))
-        project (seesaw/text (seesaw/select editor [:#project]))]
+        project (seesaw/text (seesaw/select editor [:#project]))
+        mode (if (manual? editor) :manual-config :file-config)]
     (swap! db/db* assoc-in [:settings :nrepl-host] host)
     (.setNreplHost (.getOptions configuration-base) host)
+    (swap! db/db* assoc-in [:settings :mode] mode)
+    (.setMode (.getOptions configuration-base) (name mode))
     (when-let [nrepl-port (parse-long (seesaw/text (seesaw/select editor [:#nrepl-port])))]
       (swap! db/db* assoc-in [:settings :nrepl-port] nrepl-port)
       (.setNreplPort (.getOptions configuration-base) (str nrepl-port))
@@ -66,8 +106,9 @@
   (when-let [nrepl-port (parse-long (.getNreplPort (.getOptions configuration-base)))]
     (swap! db/db* assoc-in [:settings :nrepl-port] nrepl-port))
   (when-let [project (not-empty (.getProject (.getOptions configuration-base)))]
-    (swap! db/db* assoc-in [:settings :project] project)))
-(set! *warn-on-reflection* true)
+    (swap! db/db* assoc-in [:settings :project] project))
+  (when-let [mode (not-empty (.getMode (.getOptions configuration-base)))]
+    (swap! db/db* assoc-in [:settings :mode] (keyword mode))))
 
 (defn ^:private reset-editor-from-settings []
   (let [editor @config.factory.base/editor-view*