+++ src/main/clojure/com/github/clojure_repl/intellij/db.clj
@@ -3,7 +3,8 @@
 (defonce db* (atom {:current-nrepl {:session-id nil
                                     :ns "user"}
                     :settings {:nrepl-port nil
-                               :nrepl-host "localhost"}
+                               :nrepl-host "localhost"
+                               :remote-repl-mode :manual-config}
                     :on-repl-file-loaded-fns []
                     :on-repl-evaluated-fns []
                     :ops {}}))