+++ src/main/kotlin/run-configuration-options.kt
@@ -7,6 +7,7 @@ class ReplRemoteRunOptions : RunConfigurationOptions() {
     private var nreplHostOption = string("").provideDelegate(this, "nreplHost")
     private var nreplPortOption = string("").provideDelegate(this, "nreplPort")
     private var projectOption = string("").provideDelegate(this, "project")
+    private var modeOption = string("").provideDelegate(this, "mode")
 
     var nreplHost: String
         get() = nreplHostOption.getValue(this) ?: ""
@@ -19,6 +20,10 @@ class ReplRemoteRunOptions : RunConfigurationOptions() {
     var project: String
         get() = projectOption.getValue(this) ?: ""
         set(value) = projectOption.setValue(this, value)
+
+    var mode: String
+        get() = modeOption.getValue(this) ?: ""
+        set(value) = modeOption.setValue(this, value)
 }
 
 class ReplLocalRunOptions : RunConfigurationOptions() {