+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add support for starting REPL from inside IntelliJ. #40
+- Add support to read .nrepl-port file to connect to a running nREPL process. #5
 
 ## 0.1.3